mock use user store, renaming store for just users
