+++ services/ui-src/.env_example
@@ -11,6 +11,7 @@ COGNITO_USER_POOL_CLIENT_DOMAIN=main-login-21guh5mqnb6otpoli8o9irlj94.auth.us-ea
 COGNITO_USER_POOL_ID=us-east-1_z8B48Bul1
 LD_PROJECT_KEY=mdct-mfp
 LOCAL_LOGIN=true
+POST_SIGNOUT_REDIRECT=https://test.home.idm.cms.gov/app/UserHome
 REACT_APP_LD_SDK_CLIENT=645519413f03ed59dca0de7b #pragma allowlist secret
 S3_ATTACHMENTS_BUCKET_NAME=local-uploads
 S3_ATTACHMENTS_BUCKET_REGION=us-east-1
