+++ services/ui-src/src/types/index.ts
@@ -3,3 +3,4 @@ export * from "./formFields";
 export * from "./other";
 export * from "./reports";
 export * from "./users";
+export * from "./states";
