+++ services/ui-src/src/utils/auth/UserProvider.tsx
@@ -8,7 +8,7 @@ import {
 import { useLocation } from "react-router-dom";
 import { Auth } from "aws-amplify";
 import config from "config";
-import { initAuthManager, updateTimeout, getExpiration, useStore } from "utils";
+import { initAuthManager, updateTimeout, getExpiration, useUser } from "utils";
 import { PRODUCTION_HOST_DOMAIN } from "../../constants";
 
 import { MFPUser, UserContextShape, UserRoles } from "types/users";
@@ -32,7 +32,7 @@ export const UserProvider = ({ children }: Props) => {
     setShowLocalLogins,
     logout,
     loginWithIDM,
-  } = useStore();
+  } = useUser();
 
   // initialize the authentication manager that oversees timeouts
   initAuthManager();
